+++ user/rust/src/syscall.rs
@@ -78,6 +78,6 @@ pub fn sys_exec(path: *const u8) -> isize {
     syscall(SYSCALL_EXEC,path as usize,0,0)
 }
 
-pub fn sys_wait(waitpid: usize, xstate: *mut usize) -> isize {
+pub fn sys_wait(waitpid: usize, xstate: *mut i32) -> isize {
     syscall(SYSCALL_WAIT, waitpid, xstate as usize, 0)
 }
